+++ rollup.config.js
@@ -43,7 +43,10 @@ export default {
       "styled-components": path.resolve(__dirname, "node_modules", "styled-components")
     }),
     babel({
-      exclude: "node_modules/**"
+      exclude: "node_modules/**",
+      presets: [["@babel/preset-env", { modules: false, loose: true }], "@babel/preset-react"],
+      plugins: [["babel-plugin-styled-components", "transform-runtime"]],
+      babelrc: false
     }),
     commonjs()
   ]